+++ src/Config/config.php
@@ -10,6 +10,7 @@ return [
     'cookie_name' => env('WA_COOKIE_NAME', 'waccount_token'),
     'login_route' => env('WA_LOGIN_ROUTE', 'login'),
     'home_route' => env('WA_HOME_ROUTE', 'home'),
+    'logout_route' => env('WA_LOGOUT_ROUTE', 'logout'),
 
     // API
     'api' => [
@@ -38,11 +39,11 @@ return [
     'routes' => [
         'api' => [
             'enabled' => env('WA_ROUTES_API_ENABLED', false),
-            'prefix' => env('WA_ROUTES_API_PREFIX', 'waccount'),
+            'prefix' => env('WA_ROUTES_API_PREFIX', 'waccount.api'),
         ],
         'web' => [
             'enabled' => env('WA_ROUTES_WEB_ENABLED', false),
-            'prefix' => env('WA_ROUTES_WEB_PREFIX', 'waccount'),
+            'prefix' => env('WA_ROUTES_WEB_PREFIX', 'waccount.web'),
         ],
     ],