+++ src/Routes/api.php
@@ -2,7 +2,7 @@
 
 use Illuminate\Support\Facades\Route;
 use Werify\Account\Laravel\Http\Controllers\V1\Auth\Classic\ApiController;
-use Werify\Account\Laravel\Http\Controllers\V1\Authorize\ClassicController as ClassicControllerAlias;
+use Werify\Account\Laravel\Http\Controllers\V1\Authorize\Classic\ApiController as ClassicControllerAlias;
 
 Route::prefix(config('waccount.routes.api.prefix'))->group(function () {